Performance: avoid unnecessary hosting provider detection when possible. (#41890)
* Status: add parameter to bring more specificity to host detection

In a lot of situations, we are not interested in knowing the exact hosting provider ; we just want to know if the site is on the WordPress.com platform (and if so, which one).

* Script data: provide less specific host info for Jetpack Social

We do not need to know exactly what host we're with. It's enough to know it's a WordPress.com platform.

See #41076 where this was set up.

* Boost: be more specific in hosting provider check.

We only need to check for specific WordPress.com platforms here (WoA or Atomic), so we can ditch the generic host detection and only look for those 2.

This was added in #39155.

* My Jetpack: remove unused parameter.

This was added in #34864, but never actually used. We can consequently remove it.

* Update Phan baseline

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/13432183981

Upstream-Ref: Automattic/jetpack@e3ea28e5ab5ba72722385592010cfa4ddb3f9d56